requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.8.1
asyncio==3.4.3
attrs==25.3.0
billiard==4.2.1
celery==5.5.1
certifi==2025.6.15
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.2
distlib==0.3.9
dj-rest-auth==7.0.1
Django==5.2.3
django-allauth==65.9.0
django-ckeditor==6.7.3
django-cors-headers==4.7.0
django-environ==0.12.0
django-etc==1.4.0
django-filter==24.3
django-hitcount==1.3.5
django-js-asset==3.1.2
django-modeltranslation==0.19.13
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
docutils==0.21.2
drf-yasg==1.21.10
filelock==3.18.0
flower==2.0.1
frozenlist==1.7.0
greenlet==3.2.3
gunicorn==23.0.0
humanize==4.12.3
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
kombu==5.5.4
MarkupSafe==3.0.2
msgpack==1.1.1
multidict==6.6.3
numpy==2.3.1
packaging==25.0
phonenumbers==9.0.9
pillow==11.2.1
pipenv==2025.0.3
platformdirs==4.3.8
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-environ==0.4.54
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.4
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.41
sqlparse==0.5.3
tblib==3.1.0
tornado==6.5.1
twilio==9.6.5
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
virtualenv==20.31.2
wcwidth==0.2.13
wheel==0.45.1
yarl==1.20.1
